Fixed money precision in some cases
When a product price was set as 9.87, then the value was saved as 9.86. This
problem relies in how Symfony Forms plays with numbers, even when are decimals.
We've solved this by giving the responsibility of converting floats to integers
and viceversa to the MoneyDataMaper, so we can safely do that conversion.